+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add Markdown wizard in LTI context
+- Install cunningham
+- Use cunningham to replace markdown buttons
 
 ### Changed